% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{tnt_num_docs}
\alias{tnt_num_docs}
\title{Number of searchable documents}
\usage{
tnt_num_docs(idx)
}
\arguments{
\item{idx}{A \code{tnt_index}.}
}
\value{
A single numeric value (committed document count).
}
\description{
Number of searchable documents
}
\examples{
idx <- tnt_index_df(data.frame(t = "hello world"), text = t)
tnt_num_docs(idx)
}
