% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{tnt_stemmers}
\alias{tnt_stemmers}
\title{List supported stemmer languages}
\usage{
tnt_stemmers()
}
\value{
A character vector of language names accepted by the \code{stemmer}
argument of \code{\link[=tnt_text]{tnt_text()}}.
}
\description{
List supported stemmer languages
}
\examples{
tnt_stemmers()
}
