% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tantivyr-package.R
\docType{package}
\name{tantivyr-package}
\alias{tantivyr}
\alias{tantivyr-package}
\title{tantivyr: Fast Full-Text Search for R with 'Tantivy'}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Index data frames and document collections and run fast full-text search entirely on your machine. 'tantivyr' wraps the 'Tantivy' Rust search engine (a 'Lucene'-inspired library) to provide 'BM25' ranking, structured filters, snippet highlighting and incremental updates over an on-disk or in-memory index. First-class support is provided for stemming and stop words in Portuguese and English, making it well suited to public documents, news clippings, extracted 'PDF' text, transcripts and legal acts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://strategicprojects.github.io/tantivyr/}
  \item \url{https://github.com/StrategicProjects/tantivyr}
  \item Report bugs at \url{https://github.com/StrategicProjects/tantivyr/issues}
}

}
\author{
\strong{Maintainer}: André Leite \email{leite@de.ufpe.br}

Authors:
\itemize{
  \item André Leite \email{leite@de.ufpe.br}
}

Other contributors:
\itemize{
  \item StrategicProjects [copyright holder, funder]
}

}
\keyword{internal}
